+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ### Bugfixes
 
 - [#44](https://github.com/lodastack/agent/pull/44): Random send report/ns data to server
+- [#46](https://github.com/lodastack/agent/pull/46): filter link local multicast addresses
 
 ## v0.1.0 [2017-03-10]